# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dark Thila Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionStatusStatus =
  (typeof SessionStatusStatus)[keyof typeof SessionStatusStatus];

export const SessionStatusStatus = {
  idle: "idle",
  qr: "qr",
  pairing: "pairing",
  connected: "connected",
  reconnecting: "reconnecting",
  disconnected: "disconnected",
} as const;

export interface SessionStatus {
  sessionId: string;
  phoneNumber: string;
  status: SessionStatusStatus;
  qrCode?: string | null;
  pairingCode?: string | null;
  owner?: string | null;
  logo?: string | null;
  retryCount?: number | null;
}

export type ConnectRequestMethod =
  (typeof ConnectRequestMethod)[keyof typeof ConnectRequestMethod];

export const ConnectRequestMethod = {
  qr: "qr",
  pairing: "pairing",
} as const;

export interface ConnectRequest {
  sessionId: string;
  phoneNumber: string;
  method: ConnectRequestMethod;
}

export interface ConnectResponse {
  success: boolean;
  message: string;
  sessionId: string;
}

export interface DisconnectRequest {
  sessionId: string;
}

export interface DisconnectResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}