# lib/api-zod/src/generated/types/disconnectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dark Thila Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DisconnectRequest {
  sessionId: string;
}